move ProductsBatchLoader::loadByEntities() from project-base to the frontend-api package
