+++ project-base/app/config/packages/overblog_dataloader.yaml
@@ -34,7 +34,7 @@ overblog_dataloader:
             batch_load_fn: "@Shopsys\\LuigisBoxBundle\\Model\\Batch\\LuigisBoxBatchLoader:loadByBatchData"
         productsByEntitiesBatchLoader:
             alias: "products_by_entities_batch_loader"
-            batch_load_fn: "@App\\FrontendApi\\Model\\Product\\ProductsBatchLoader:loadByEntities"
+            batch_load_fn: "@Shopsys\\FrontendApiBundle\\Model\\Product\\BatchLoad\\ProductsBatchLoader:loadByEntities"
         productsSellableByIdsBatchLoader:
             alias: "products_sellable_by_ids_batch_loader"
             batch_load_fn: "@Shopsys\\FrontendApiBundle\\Model\\Product\\BatchLoad\\ProductsBatchLoader:loadSellableByIds"
