+++ project-base/app/config/services.yaml
@@ -420,9 +420,6 @@ services:
     Shopsys\FrameworkBundle\Model\Product\ProductDataFactory:
         alias: App\Model\Product\ProductDataFactory
 
-    Shopsys\FrameworkBundle\Model\Product\ProductElasticsearchProvider:
-        alias: App\Model\Product\ProductElasticsearchProvider
-
     Shopsys\FrameworkBundle\Model\Product\ProductFacade:
         alias: App\Model\Product\ProductFacade
 
@@ -479,9 +476,6 @@ services:
 
     Shopsys\FrontendApiBundle\Model\Mutation\Customer\DeliveryAddress\DeliveryAddressMutation:
         alias: App\FrontendApi\Mutation\Customer\DeliveryAddressMutation
-
-    Shopsys\FrontendApiBundle\Model\Product\BatchLoad\ProductsBatchLoader:
-        alias: App\FrontendApi\Model\Product\ProductsBatchLoader
 
     Shopsys\FrontendApiBundle\Model\Token\TokenAuthenticator:
         alias: App\FrontendApi\Model\Token\TokenAuthenticator
